refactor(dashboard): move FilterSelect from Audit.tsx to new filter-select.tsx file
