+++ platform/dashboard/src/components/audit/pieces.tsx
@@ -23,13 +23,7 @@ import type {
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Card } from "@/components/ui/card";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
+import { FilterSelect } from "@/components/ui/filter-select";
 import {
   Table,
   TableBody,
@@ -56,40 +50,6 @@ const toDatum = (r: AuditBreakdownRow): BreakdownDatum => ({
 
 // Radix Select items can't carry value="" — UI-local stand-in for "all".
 const ALL = "__all__";
-
-// Module-level (not re-created per render). Exported — reused by the
-// Usage page's filter bar (components/usage/pieces.tsx), which needs the
-// same "select with an all-option" shape for agent/model/user.
-export function FilterSelect({
-  value,
-  all,
-  opts,
-  onChange,
-}: {
-  value: string;
-  all: string;
-  opts: string[];
-  onChange: (v: string) => void;
-}) {
-  return (
-    <Select
-      value={value || ALL}
-      onValueChange={(v) => onChange(v === ALL ? "" : v)}
-    >
-      <SelectTrigger className="h-8 w-auto min-w-32 gap-1.5 text-[13px]">
-        <SelectValue />
-      </SelectTrigger>
-      <SelectContent>
-        <SelectItem value={ALL}>{all}</SelectItem>
-        {opts.map((o) => (
-          <SelectItem key={o} value={o} className="font-mono text-xs">
-            {o}
-          </SelectItem>
-        ))}
-      </SelectContent>
-    </Select>
-  );
-}
 
 // ————————————————————————————————————————————— Filter bar
 export function FilterBar({
